Move badge and active functions into one file each
