+++ signinapp/templates/badges.html.jinja2
@@ -3,52 +3,52 @@
   Badge List
 {% endblock title %}
 {% block content %}
-  <div class="pt-3 container">
-    <div class="d-flex flex-row align-items-center justify-content-between">
-      <h1>Available Badges!</h1>
-      <div>
-        {%- if current_user.is_authenticated and current_user.role.admin -%}
-          <a class="btn btn-primary me-1"
-             href="{{ url_for('admin.new_badge')}}"
-             role="button">New Badge</a>
-        {%- endif -%}
-        <a class="btn btn-primary"
-           href="{{ url_for('search.badges')}}"
-           role="button">Search Badges</a>
-      </div>
-    </div>
-    {{- render_messages() -}}
-    <div class="table-responsive">
-      <table id="badges">
-        <thead>
-          <tr>
-            <th scope="col">Badge</th>
-            <th scope="col">ID</th>
-            <th scope="col">Description</th>
-            <th scope="col">Tools</th>
-          </tr>
-        </thead>
-        <tbody>
-          {%- for badge in badges -%}
-            <tr>
-              <th>
-                <a class="link-light"
-                   href="{{ url_for('user.badge', badge_id=badge.id)}}">{{ show_badge(badge)}}</a>
-              </th>
-              <td>{{ badge.id }}</td>
-              <td>{{ badge.description }}</td>
-              <td>
-                <a class="btn btn-sm btn-secondary"
-                   href="{{ url_for('admin.edit_badge', badge_id=badge.id)}}"
-                   role="button">Edit</a>
-                <a class="btn btn-sm btn-secondary"
-                   href="{{ url_for('mentor.award_badge', badge_id=badge.id)}}"
-                   role="button">Award</a>
-              </td>
-            </tr>
-          {%- endfor -%}
-        </tbody>
-      </table>
+<div class="pt-3 container">
+  <div class="d-flex flex-row align-items-center justify-content-between">
+    <h1>Available Badges!</h1>
+    <div>
+      {%- if current_user.is_authenticated and current_user.role.admin -%}
+      <a class="btn btn-primary me-1"
+        href="{{url_for('badge.new')}}"
+        role="button">New Badge</a>
+      {%- endif -%}
+      <a class="btn btn-primary"
+        href="{{ url_for('badge.search')}}"
+        role="button">Search Badges</a>
     </div>
   </div>
+  {{- render_messages() -}}
+  <div class="table-responsive">
+    <table id="badges">
+      <thead>
+        <tr>
+          <th scope="col">Badge</th>
+          <th scope="col">ID</th>
+          <th scope="col">Description</th>
+          <th scope="col">Tools</th>
+        </tr>
+      </thead>
+      <tbody>
+        {%- for badge in badges -%}
+          <tr>
+            <th>
+              <a class="link-light"
+                 href="{{ url_for('badge.view', badge_id=badge.id)}}">{{ show_badge(badge) }}</a>
+            </th>
+            <td>{{badge.id}}</td>
+            <td>{{badge.description}}</td>
+            <td>
+              <a class="btn btn-sm btn-secondary"
+                  href="{{ url_for('badge.edit', badge_id=badge.id)}}"
+                  role="button">Edit</a>
+              <a class="btn btn-sm btn-secondary"
+                  href="{{ url_for('badge.award', badge_id=badge.id)}}"
+                  role="button">Award</a>
+            </td>
+          </tr>
+        {%- endfor -%}
+      </tbody>
+    </table>
+  </div>
+</div>
 {% endblock content %}
