+++ signinapp/templates/base.html.jinja2
@@ -90,10 +90,10 @@
                   <ul class="dropdown-menu dropdown-menu-dark"
                       aria-labelledby="navbarSearchDropdown">
                     <li>
-                      <a class="dropdown-item" href="{{ url_for('mentor.active')}}">Active Stamps</a>
+                      <a class="dropdown-item" href="{{ url_for('active.get')}}">Active Stamps</a>
                     </li>
                     <li>
-                      <a class="dropdown-item" href="{{ url_for('mentor.all_badges')}}">Badges</a>
+                      <a class="dropdown-item" href="{{ url_for('badge.all')}}">Badges</a>
                     </li>
                     <li>
                       <a class="dropdown-item" href="{{ url_for('search.hours')}}">Search Hours</a>
