+++ signinapp/user.py
@@ -1,9 +1,9 @@
-from flask import Blueprint, Flask, redirect, request, url_for, flash
+from flask import Blueprint, Flask, redirect, url_for, flash
 from flask.templating import render_template
 from flask_login import current_user, login_required
 from sqlalchemy.future import select
 
-from .model import Badge, BadgeAward, EventType, User, db
+from .model import EventType, User, db
 
 user = Blueprint("user", __name__)
 
@@ -26,19 +26,6 @@ def profile(username):
         return redirect(url_for("index"))
     event_types = db.session.scalars(select(EventType))
     return render_template("profile.html.jinja2", user=user, event_types=event_types)
-
-
-@user.route("/badge")
-@login_required
-def badge():
-    if bid := request.args.get("badge_id"):
-        bid = int(bid)
-        badge: Badge = db.session.get(Badge, bid)
-        awards: list[BadgeAward] = sorted(
-            [a for a in badge.awards], key=lambda u: u.owner.name
-        )
-        return render_template("badge.html.jinja2", badge=badge, awards=awards)
-    return redirect(url_for("mentor.all_badges", badge_id=badge.id))
 
 
 def init_app(app: Flask):
